Fix interruption of standard call by progressive call without breaking page
This bug showed itself on reletivly slow servers but it was also possible to reproduce it on fast servers, by calling a page with slow load (like System Time or Dashboard), and without waiting for completion, call for progressive page, like Re-Check System Configuration. The result actually was quite funny to watch. Give it a try.