Fix re-authentication pop up screen locking OS system shortcuts
* Note: This patch simply prevents macOS SKE from kicking-in.

* Explanation: This macOS feature blocks every global event-tap from intercepting key events, which is Apple’s defense against key-loggers, affecting only the focused elements and only for password input fields to ensure that only this password field can continue to receive keystroke input.

So, what we do is just remove focus from the password input field in the background to allow macOS system hotkeys to continue to work without the need to go to the browser immediately to take an action.

Anyone wanting to understand how it works can just open the panel in Chrome or Safari, then Cmd + Click on the logout link in the navigation menu and quickly switch away to some other app. When the re-authentication screen pops up in the background, all global event-taps will be locked with the Authentic theme version 24.02 and earlier. You can test this with apps like Alfred installed which often used to switch between different desktop apps using shortcuts.